+++ aiidalab_qe/app/parameters/qeapp.yaml
@@ -8,9 +8,10 @@ run_bands: false
 run_pdos: false
 
 ## Codes
-dos_code: dos-7.2@localhost
-projwfc_code: projwfc-7.2@localhost
-pw_code: pw-7.2@localhost
+codes:
+    dos_code: dos-7.2@localhost
+    projwfc_code: projwfc-7.2@localhost
+    pw_code: pw-7.2@localhost
 
 
 basic:
