+++ setup.cfg
@@ -50,6 +50,11 @@ dev =
 [options.package_data]
 aiidalab_qe.parameters = qeapp.yaml
 
+[options.entry_points]
+aiidalab_qe.property =
+    bands = aiidalab_qe.app.plugins.bands:property
+    pdos = aiidalab_qe.app.plugins.pdos:property
+
 [aiidalab]
 title = Quantum ESPRESSO
 description = Perform Quantum ESPRESSO calculations
