+++ tests/test_configure_advanced_settings.py
