+++ tests/test_configure_basic_settings.py
